Move WizardPanelsRouter to WizardPanels/utils
Enact-DCO-1.0-Signed-off-by: Juwon Jeong (juwon.jeong@lge.com)
